+++ transforms/code/code_quality/kfp_ray/code_quality_wf.py
@@ -21,16 +21,23 @@ from workflow_support.compile_utils import (
     ComponentUtils,
 )
 
+from python_apiserver_client.params import (
+        EnvironmentVariables,
+        EnvVarFrom,
+        EnvVarSource,
+)
 
 # the name of the job script
 EXEC_SCRIPT_NAME: str = "code_quality_transform_ray.py"
 PREFIX: str = ""
 
 task_image = "quay.io/dataprep1/data-prep-kit/code_quality-ray:latest"
-# Need to further investigate but for now:
-# When using None for the hf_token, the workflow works fine with v1 
-# but for v2, it fails with None and requires the string "None"
-HF_TOKEN= "None" if os.getenv("KFPv2", "0") == "1" else None
+
+# The name of the secret that holds the HugginFace token
+HF_SECRET = "hf-secret"
+# The secret key that holds the HugginFace token
+HF_SECRET_KEY = "hf-token"
+
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:latest"
 
@@ -52,7 +59,6 @@ def compute_exec_params_func(
     cq_contents_column_name: str,
     cq_language_column_name: str,
     cq_tokenizer: str,
-    cq_hf_token: str,
 ) -> dict:
     from runtime_utils import KFPUtils
 
@@ -68,7 +74,6 @@ def compute_exec_params_func(
         "cq_contents_column_name": cq_contents_column_name,
         "cq_language_column_name": cq_language_column_name,
         "cq_tokenizer": cq_tokenizer,
-        "cq_hf_token": cq_hf_token,
     }
 
 
@@ -94,6 +99,13 @@ cleanup_ray_op = comp.load_component_from_file(component_spec_path + "deleteRayC
 # Task name is part of the pipeline name, the ray cluster name and the job name in DMF.
 TASK_NAME: str = "code_quality"
 
+# HuggingFace token is exported as environment variables in Ray node pods.
+# Alternatively, the secret name can be passed to the KFP component,
+# which will set it as an environment variable in the Ray nodes.
+# In this option the secret name can be set at runtime
+# but is dependent on the KFP version.
+env_v = EnvVarFrom(source=EnvVarSource.SECRET, name=HF_SECRET, key=HF_SECRET_KEY)
+envs = EnvironmentVariables(from_ref={"HF_READ_ACCESS_TOKEN": env_v})
 
 # Pipeline to invoke execution on remote resource
 @dsl.pipeline(
@@ -105,8 +117,8 @@ def code_quality(
     ray_name: str = "code_quality-kfp-ray",  # name of Ray cluster
     ray_run_id_KFPv2: str = "",   # Ray cluster unique ID used only in KFP v2
     # Add image_pull_secret and image_pull_policy to ray workers if needed
-    ray_head_options: dict = {"cpu": 1, "memory": 4, "image": task_image},
-    ray_worker_options: dict = {"replicas": 2, "max_replicas": 2, "min_replicas": 2, "cpu": 2, "memory": 4, "image": task_image},
+    ray_head_options: dict = {"cpu": 1, "memory": 4, "image": task_image, "environment": envs.to_dict()},
+    ray_worker_options: dict = {"replicas": 2, "max_replicas": 2, "min_replicas": 2, "cpu": 2, "memory": 4, "image": task_image, "environment": envs.to_dict()},
     server_url: str = "http://kuberay-apiserver-service.kuberay.svc.cluster.local:8888",
     # data access
     data_s3_config: str = "{'input_folder': 'test/code_quality/input/', 'output_folder': 'test/code_quality/output/'}",
@@ -121,7 +133,6 @@ def code_quality(
     cq_contents_column_name: str = "contents",
     cq_language_column_name: str = "language",
     cq_tokenizer: str = "codeparrot/codeparrot",
-    cq_hf_token: str = HF_TOKEN,
     # additional parameters
     additional_params: str = '{"wait_interval": 2, "wait_cluster_ready_tmout": 400, "wait_cluster_up_tmout": 300, "wait_job_ready_tmout": 400, "wait_print_tmout": 30, "http_retries": 5, "delete_cluster_delay_minutes": 0}',
 ):
@@ -161,7 +172,6 @@ def code_quality(
     :param cq_contents_column_name - Name of the column holds the data to process
     :param cq_language_column_name - Name of the column holds the programming language details
     :param cq_tokenizer - Name or path to the tokenizer
-    :param cq_hf_token - Huggingface auth token to download and use the tokenizer
     :return: None
     """
     # In KFPv2 dsl.RUN_ID_PLACEHOLDER is deprecated and cannot be used since SDK 2.5.0. On another hand we cannot create
@@ -192,7 +202,6 @@ def code_quality(
             cq_contents_column_name=cq_contents_column_name,
             cq_language_column_name=cq_language_column_name,
             cq_tokenizer=cq_tokenizer,
-            cq_hf_token=cq_hf_token,
         )
 
         ComponentUtils.add_settings_to_component(compute_exec_params, ONE_HOUR_SEC * 2)