+++ transforms/language/text_encoder/kfp_ray/text_encoder_wf.py
@@ -28,7 +28,7 @@ task_image = "quay.io/dataprep1/data-prep-kit/text_encoder-ray:latest"
 EXEC_SCRIPT_NAME: str = "-m dpk_text_encoder.ray.transform"
 
 # components
-base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:0.2.3"
+base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:latest"
 
 # path to kfp component specifications files
 component_spec_path = os.getenv("KFP_COMPONENT_SPEC_PATH", DEFAULT_KFP_COMPONENT_SPEC_PATH)